Got attacking to work. (Next htings to work on clarified in description)
My issue was I didn't have player 2 attack after player 1, so then it would only have the player with the higher attack speed attack always.

Work on fluidity of the attacks. Next up is attribute modification, rock paper scissors with type, and then reworking attacking to have moves involved, may need subclasses for this? Not sure but it'll definitely be interesting to work on